refactor(profile): move profile data models to data layer
Moves `UserProfile`, `AccessoryItem`, `AccessorySlot`, `Rarity`, `AccentVariant`, and `UserStats` data classes from the `ui/profile` package to the central `data` package.

This refactoring consolidates domain models into a common data layer, improving code organization and separation of concerns. The `UserStats.kt` file was deleted and its contents were merged into the new `ProfileStats.kt` file.
